requirements.txt
web3==5.24.0
python-dotenv==0.19.2